Relicense to MIT License.

+++ src/MemoryPool.cpp
@@ -1,20 +1,7 @@
-/*
- *  This file is part of ICon7.
- *  Copyright (C) 2024-2025 Marek Zalewski aka Drwalin
- *
- *  ICon7 is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  ICon7 is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
+// Copyright (C) 2024-2025 Marek Zalewski aka Drwalin
+// 
+// This file is part of ICon7 project under MIT License
+// You should have received a copy of the MIT License along with this program.
 
 #include "../include/icon7/Debug.hpp"
 
@@ -41,7 +28,7 @@ void MemoryPool::PrintStats()
 #endif
 }
 
-AllocatedObject<void> MemoryPool::Allocate(uint64_t bytes)
+AllocatedObject<void> MemoryPool::Allocate(size_t bytes)
 {
 #if ICON7_USE_RPMALLOC
 	static int ____staticInitilizeRpmalloc = (rpmalloc_initialize(), 0);
@@ -59,7 +46,7 @@ AllocatedObject<void> MemoryPool::Allocate(uint64_t bytes)
 #endif
 }
 
-void MemoryPool::Release(void *ptr, uint64_t bytes)
+void MemoryPool::Release(void *ptr, size_t bytes)
 {
 #if ICON7_USE_RPMALLOC
 	rpfree(ptr);
